// packages/shared/src/enums/WeaponType.ts
/** Available weapon types. */
export enum WeaponType {
  FISTS = 0,
  DAGGER = 1,
  SHORT_SWORD = 2,
  LONG_SWORD = 3,
  HAMMER = 4,
  LARGE_AXE = 5,
  BLADED_AXE = 6,
  DOUBLE_AXE = 7,
  SPEAR = 8,
  POLEARM = 9,
  STAFF = 10,
  THROWING_AXE = 11,
  SHORT_BOW = 12,
  CROSSBOW = 13,
  SMALL_SHIELD = 14,
  LARGE_SHIELD = 15,
}